+++ lib/transformer.dart
@@ -3,6 +3,8 @@
  * Thanks to juha.komulainen@evident.fi for inspiration and some code
  * (Copyright (c) 2013 Evident Solutions Oy) from package http://pub.dartlang.org/packages/sass
  *
+ * less_dart v 0.2.1  20150321 .html, * ! in entry_points
+ * less_dart v 0.2.1  20150317 https://github.com/luisvt - AggregateTransform
  * less_dart v 0.1.4  20150112 'build_mode: dart' as default
  * less_dart v 0.1.0  20141230
  * less_node v 0.2.1  20141212 niceDuration, other_flags argument
@@ -15,90 +17,86 @@
 library less.transformer;
 
 import 'dart:async';
-import 'dart:math';
-import 'dart:io';
-import 'less.dart';
-import 'package:barback/barback.dart';
 
-const String INFO_TEXT = '[Info from less-dart]';
-const String BUILD_MODE_LESS = 'less';
-const String BUILD_MODE_DART = 'dart';
-const String BUILD_MODE_MIXED = 'mixed';
+import 'srcTransformer/base_transformer.dart';
+
+import 'package:barback/barback.dart';
 
 /*
  * Transformer used by 'pub build' & 'pub serve' to convert .less files to .css
- * Based on lessc over nodejs executing a process like
- * CMD> lessc --flags input.less output.css
- * It use one or various files as entry point and produces the css files
- * To mix several .less files in one, the input contents could be "@import 'filexx.less'; ..." directives
+ * Also works in .html files, converting <less> tags to <style>
+ * entry_points has default values and support * in path, and exclusion paths (!).
  * See http://lesscss.org/ for more information
  */
-class LessTransformer extends Transformer {
+class FileTransformer extends AggregateTransformer {
   final BarbackSettings settings;
   final TransformerOptions options;
 
-  bool get isBuildModeLess => options.build_mode == BUILD_MODE_LESS;   //input file, output file
-  bool get isBuildModeDart => options.build_mode == BUILD_MODE_DART;   //input stdin, output stdout
-  bool get isBuildModeMixed => options.build_mode == BUILD_MODE_MIXED; //input file, output stdout
+  EntryPoints entryPoints;
 
-  LessTransformer(BarbackSettings settings):
+  FileTransformer(BarbackSettings settings):
     settings = settings,
-    options = new TransformerOptions.parse(settings.configuration);
+    options = new TransformerOptions.parse(settings.configuration) {
 
-  LessTransformer.asPlugin(BarbackSettings settings):
+    entryPoints = new EntryPoints();
+    entryPoints.addPaths(options.entry_points);
+    entryPoints.assureDefault(['*.less', '*.html']);
+  }
+
+  FileTransformer.asPlugin(BarbackSettings settings):
     this(settings);
 
-  Future<bool> isPrimary (AssetId id) {
-    return new Future.value(_isEntryPoint(id));
+  @override
+  classifyPrimary(AssetId id) {
+    if (entryPoints.check(id.path)) return id.extension.toLowerCase();
+    return null;
   }
 
-  Future apply(Transform transform) {
-    List<String> flags = _createFlags();  //to build process arguments
-    var id = transform.primaryInput.id;
-    String inputFile = id.path;
-    String outputFile = getOutputFileName(id);
+  @override
+  Future apply(AggregateTransform transform) {
+    return transform.primaryInputs.toList().then((assets) {
+      return Future.wait(assets.map((asset) {
+        return asset.readAsString().then((content) {
+          List<String> flags = _createFlags();  //to build process arguments
+          var id = asset.id;
 
-    switch (options.build_mode) {
-      case BUILD_MODE_DART:
-        flags.add('-');
-        break;
-      case BUILD_MODE_MIXED:
-        flags.add(inputFile);
-        break;
-      case BUILD_MODE_LESS:
-      default:
-        flags.add(inputFile);
-        flags.add(outputFile);
-    }
-
-    ProcessInfo processInfo = new ProcessInfo(options.executable, flags);
-    if (isBuildModeDart) processInfo.inputFile = inputFile;
-    if (isBuildModeMixed || isBuildModeDart) processInfo.outputFile = outputFile;
-
-    return transform.primaryInput.readAsString().then((content){
-      transform.consumePrimary();
-      return executeProcess(options.executable, flags, content, processInfo).then((output) {
-
-        if (isBuildModeMixed || isBuildModeDart){
-          transform.addOutput(new Asset.fromString(new AssetId(id.package, outputFile), output));
-        }
-      });
+          if (id.extension.toLowerCase() == '.html') {
+            HtmlTransformer htmlProcess = new HtmlTransformer(content, id.path);
+            return htmlProcess.transform(flags).then((process){
+              if (process.deliverToPipe) {
+                transform.addOutput(new Asset.fromString(new AssetId(id.package, id.path), process.outputContent));
+                if (process.isError || !options.silence) print (process.message);
+                if (process.isError) {
+                  print('**** ERROR ****  see build/' + process.outputFile + '\n');
+                  print(process.errorMessage);
+                }
+              }
+            });
+          } else if (id.extension.toLowerCase() == '.less') {
+            LessTransformer lessProcess = new LessTransformer(content, id.path,
+                getOutputFileName(id, assets.length), options.build_mode);
+            return lessProcess.transform(flags).then((process) {
+              if (process.deliverToPipe) {
+                transform.addOutput(new Asset.fromString(new AssetId(id.package, process.outputFile), process.outputContent));
+              }
+              if (process.isError || !options.silence) print (process.message);
+              if (process.isError) {
+                String resultFile = process.deliverToPipe ? ('build/' + process.outputFile) :process.outputFile;
+                print('**** ERROR ****  see ' + resultFile + '\n');
+                print(process.errorMessage);
+              }
+            });
+          }
+        });
+      }));
     });
-  }
-
-  /*
-   * Only returns true in entry_point(s) file
-   */
-  bool _isEntryPoint(AssetId id) {
-    if (id.extension != '.less') return false;
-    return (options.entry_points.contains(id.path));
   }
 
   List<String> _createFlags(){
     List<String> flags = [];
 
     flags.add('--no-color');
-    if (options.cleancss) flags.add('--clean-css');
+    //if (options.cleancss) flags.add('--clean-css');
     if (options.compress) flags.add('--compress');
     if (options.include_path != '') flags.add('--include-path=${options.include_path}');
     if (options.other_flags != null) flags.addAll(options.other_flags);
@@ -106,118 +104,17 @@ class LessTransformer extends Transformer {
     return flags;
   }
 
-  String getOutputFileName(id) {
-    if(options.entry_points.length > 1 || options.output == '') {
+  ///
+  /// For .less files returns the outputFilename
+  ///
+  /// options.output only works if we process 1 .less file
+  /// else the name is file.less -> file.css
+  ///
+  String getOutputFileName(id, int groupCount) {
+    if(groupCount > 1 || options.output == '') {
       return id.changeExtension('.css').path;
     }
     return options.output;
-  }
-
-  /*
-   * lessc process wrapper
-   */
-  Future executeProcess(String executable, List<String> flags, String content, ProcessInfo processInfo) {
-    return runZoned((){
-      final _timeInProcess = new Stopwatch();
-
-      Less less = new Less();
-      less.stdin.write(content);
-
-      _timeInProcess.start();
-      return less.transform(flags)
-        .then((exiCode){
-          _timeInProcess.stop();
-          processInfo.nicePrint(_timeInProcess.elapsed);
-          print(less.stderr.toString());
-          if(exitCode == 0){
-            return new Future.value(less.stdout.toString());
-          } else {
-            throw new LessException(less.stderr.toString());
-          }
-        });
-    },
-    zoneValues: {#id: new Random().nextInt(10000)});
-  }
-}
-/* ************************************** */
-class TransformerOptions {
-  final List<String> entry_points;  // entry_point: web/builder.less - main file to build or [file1.less, ...,fileN.less]
-  final String include_path; // include_path: /lib/lessIncludes - variable and mixims files
-  final String output;       // output: web/output.css - result file. If '' same as web/input.css
-  final bool cleancss;       // cleancss: true - compress output by using clean-css
-  final bool compress;       // compress: true - compress output by removing some whitespaces
-
-  final String executable;   // executable: lessc - command to execute lessc  - NOT USED
-  final String build_mode;   // build_mode: dart - io managed by lessc compiler (less) by (dart) or (mixed)
-  final List other_flags;    // other options in the command line
-
-  TransformerOptions({List<String> this.entry_points, String this.include_path, String this.output, bool this.cleancss, bool this.compress,
-    String this.executable, String this.build_mode, List this.other_flags});
-
-  factory TransformerOptions.parse(Map configuration){
-
-    config(key, defaultValue) {
-      var value = configuration[key];
-      return value != null ? value : defaultValue;
-    }
-
-    List<String> readStringList(value) {
-      if (value is List<String>) return value;
-      if (value is String) return [value];
-      return null;
-    }
-
-    List<String> readEntryPoints(entryPoint, entryPoints) {
-      List<String> result = [];
-      List<String> value;
-
-      value = readStringList(entryPoint);
-      if (value != null) result.addAll(value);
-
-      value = readStringList(entryPoints);
-      if (value != null) result.addAll(value);
-
-      if (result.length < 1) print('$INFO_TEXT No entry_point supplied!');
-      return result;
-    }
-
-    return new TransformerOptions (
-        entry_points: readEntryPoints(configuration['entry_point'], configuration['entry_points']),
-        include_path: config('include_path', ''),
-        output: config('output', ''),
-        cleancss: config('cleancss', false),
-        compress: config('compress', false),
-
-        executable: config('executable', 'lessc'),
-        build_mode: config('build_mode', BUILD_MODE_DART),
-        other_flags: readStringList(configuration['other_flags'])
-    );
-  }
-}
-/* ************************************** */
-class ProcessInfo {
-  String executable;
-  List<String> flags;
-  String inputFile = '';
-  String outputFile = '';
-
-  ProcessInfo(this.executable, this.flags);
-
-  nicePrint(Duration elapsed){
-    print('$INFO_TEXT command: $executable ${flags.join(' ')}');
-    if (inputFile  != '') print('$INFO_TEXT input File: $inputFile');
-    if (outputFile != '') print('$INFO_TEXT outputFile: $outputFile');
-    print ('$INFO_TEXT $executable transformation completed in ${niceDuration(elapsed)}');
-  }
-
-  /// Returns a human-friendly representation of [duration].
-  //from barback - Copyright (c) 2013, the Dart project authors
-  String niceDuration(Duration duration) {
-  var result = duration.inMinutes > 0 ? "${duration.inMinutes}:" : "";
-
-  var s = duration.inSeconds % 59;
-  var ms = (duration.inMilliseconds % 1000) ~/ 100;
-  return result + "$s.${ms}s";
   }
 }